# Improve performance of WithLabelValues(...)