@colega 2023-10-05
There's a `fmt.Errorf(msg string, args ...interface{})` call on the next line, and we're passing `vals` as one of those `args`. Since a slice can't fit into the "value" part of the interface tuple (type, value), it escapes to the heap and a pointer to that slice is stored instead.

A longer explanation after some googling about this topic: https://stackoverflow.com/questions/39492539/go-implicit-conversion-to-interface-does-memory-allocation