@bboreham 2023-10-05
In escape analysis, the compiler tries to prove that a pointer to data does not get stored somewhere such that it lives on past the return of the function it was created in.

Interfaces defeat this analysis, since the code that implements a method could come from anywhere in the program.
Escape analysis is conservative - if there is any chance of an escape, that's enough.
So, by and large, any time you see `interface` the value is going to escape.

Oleg's change makes a deliberate copy before the slice is passed as an `interface{}`, so now the compiler can prove that the original value does not escape.